fix main

+++ src/main.rs
@@ -36,7 +36,7 @@ fn write_to_csv(filepath: &str, header: &str, data: Vec<(i32, i32, i32, i32, i32
 }
 
 fn simulate_hoyo_games(num_simulations: u64) -> io::Result<()> {
-    let num_threads = num_cpus::get();
+    let num_threads = num_cpus::get() as u64;
     let num_simulations_per_thread = num_simulations / num_threads;
 
     let games = [
@@ -104,7 +104,7 @@ fn simulate_hoyo_games(num_simulations: u64) -> io::Result<()> {
 }
 
 fn simulate_other_games(num_simulations: u64) -> io::Result<()> {
-    let num_threads = num_cpus::get();
+    let num_threads = num_cpus::get() as u64;
     let num_simulations_per_thread = num_simulations / num_threads;
     let multi_progress = MultiProgress::new();
